feat(proxy) resolve DNS hostnames post-plugins
This is mainly to address the Lambda and OpenWhisk plugins not being
able to run if the specified ùpstream_url`'s hostname does not resolve,
hence resulting in an HTTP 500 error.

If in the future some plugins need to know the IP address or the
upstream Host header, the increased granularity of the future plugins
interface or the plugins API should allow for it.